configure: Ignore nullability-completeness apple clang warnings
Some of the versions of Apple Clang produces a ton of the warnings
related to the missing nullablity specifiers on the existing codebase of
ffmpeg which significantly slows down the compilation becuase of the
produced output size (especially on CI as a part of external build systems
because they usually analyze every single line of the produced stdout).

This patch turns them off when compiling for darwing targets.

And because apple's nullablity specifiers obviously are not used in the
codebase having this warning turn of just makes sense.