avfilter/vf_perspective: factor the interpolation code out so that it's usable by other filters
As mentioned here (https://ffmpeg.org/pipermail/ffmpeg-devel/2025-June/344593.html), this is the first step of a 2 steps patch: move the interpolation code out of the vf_perspective filter so that we can use it in the zoompan filter and apply scale/crop floating point precision there.

I've moved the whole " ff_filter_execute " logic in these new files since we're going to use the exact same logic in zoompan.

Signed-off-by: Quentin Renard <contact-github@asticode.com>