avfilter/f_ebur128: move true peak calculation out of main loop
Easier to read, less convoluted, and ~30% faster. Most importantly, this
avoids repeating the redundant recalculation of the true peak on every single
sample, by moving the FIND_PEAK() loop out of the main loop. (Note that
FIND_PEAK() does not depend on the current sample index at all, so there is no
reason for it to ever be recomputed here)