gcc: Don't disable '-ftree-vectorize' if gcc version is 13 or newer
This patch modifies the FFmpeg build system to allow GCC to use the
`-ftree-vectorize` flag when the compiler version is 13 or newer.
Enabling this flag can improve performance through better loop analysis
and auto-vectorization (SIMD) opportunities in modern GCC versions.

The explicit `-fno-tree-vectorize` flag was originally added in commit
973859f5230e (2009). A previous attempt to enable `-ftree-vectorize`
was made in commit cb8646af24bd (2016), but it was reverted in
fd6dbc53855f. The reason for the revert was that the inline x86 CABAC
assembly code caused compiler errors - the compiler would run out of
available registers during vectorization, making it unable to compile
some functions. There were also reports of GCC hitting internal
compiler errors, and miscompilations on x86_64.

In commit 182663a58a7a (2023), the problematic CABAC function was made
non-inline. This significantly reduces the risk of register exhaustion
caused by inlining large assembly blocks, making the vectorizer safer
to enable for other functions.

Given improvements in GCC's vectorizer and the mitigation of the
original issue, we now re-enable `-ftree-vectorize` for GCC version 13
and above.

Signed-off-by: Jiawei <jiawei@iscas.ac.cn>