avfilter/ffmpeg_filter: rate control all filter graphs
It was never reliable to detect if a filtergraph have sources, because a filter
can act as a source only after some time, for example the loop filter.

So it is better to remove the source detection entirely and always give the
scheduler an oppurtunity to stop processing.

Fixes ticket #11604.

Signed-off-by: Marton Balint <cus@passwd.hu>