avformat/tee: refactor option processing
Instead of the convoluted nested macros, use a convenience
helper function. While this makes the code slightly longer, it
is now much clearer what is happening without running the file
through a preprocessor first.

Additionally do not mess with the internals of the dictionary,
just to save two string copies.