libavcodec/utils: Ensure allocated buffer is zero-initialized
In `av_fast_padded_mallocz`, the allocated buffer's zero-initialization is not guaranteed.
This is because it calls `av_fast_malloc`, which in turn calls `fast_malloc` with `zero_realloc=0`.
Consequently, the memory returned by the underlying `av_malloc` (used within `fast_malloc`)
is not guaranteed to be zero-initialized.

Furthermore, if `*size` is adjusted to be greater than `min_size + AV_INPUT_BUFFER_PADDING_SIZE`,
the subsequent `memset` operation will not cover the entire allocated buffer,
leaving a portion of it uninitialized.

To ensure the entire allocated buffer is properly zero-initialized, we should use `FFMAX`
to adjust the `memset` range.

Signed-off-by: xjdeng <micro6947@gmail.com>