avcodec/utvideodec: fix leaks in decode_plane() and decode_plane10()
In decode_plane() and decode_plane10(), both of these two functions use
build_buff() which allocates memory in vlc and multi. And both of them
forget to release vlc and multi when build_buff report a symbol to fill
slices with. Add cleanup label and goto cleanup first before return 0.

Signed-off-by: Lidong Yan <502024330056@smail.nju.edu.cn>