avfilter: add scale_d3d11 filter
> This part is still not acceptable. You blindly set a flag based on
> what you need, if special flags on the context are needed they should
> be negotiated between the different components, guided by the users
> use-case and input, and not hardcoded based on an arbitrary condition
> of a pixel format.

Couldn't find a way to negotiate between decoder and encoder, but it should
be possible to use scale_d3d11 filter as passthrough from which the output
can be sent directly to the encoder. So removed the change from dxva2.c

> A potentially infinite loop in get_buffer is an absolute no-go. There
> is no way to control this behavior, and the surface may never become
> available if the caller is blocked in a loop here. You assume a
> certain threading model which may not exist.
> This sort of behavior should be handled by the caller, not here (and
> no, a maximum wait time or anything like that does not resolve this
> concern, move it to another layer where the caller can control this).

Thanks for the feedback. I did try moving the wait logic out of get_buffer,
but couldn't find a clean way to handle it at a different layer. Without the loop,
it ends up breaking the pipeline with a “static surface pool size exceeded”
error, as all surfaces remain busy.

I also attempted returning EAGAIN instead of ENOMEM, but that leads to an
assertion failure in the decoder path - apparently EAGAIN isn’t expected at that
stage and causes a hard abort.

At this point I’m honestly not sure how to proceed. I'd really appreciate any
guidance or suggestions on alternative approaches that might work here.
Sorry if I’m missing something obvious.

- Dash