avfilter/vf_blackdetect: add AVX2 SIMD version
Requested by a user. Even with autovectorization enabled, the compiler
performs a quite poor job of optimizing this function, due to not being
able to take advantage of the pmaxub + pcmpeqb trick for counting the number
of pixels less than or equal-to a threshold.

blackdetect8_c:                                       4627.2 ( 1.00x)
blackdetect8_avx2:                                     163.9 (28.24x)
blackdetect16_c:                                      2474.9 ( 1.00x)
blackdetect16_avx2:                                    155.9 (15.88x)