avcodec/h264dec: implement export of video coding info for H.264
Hooks into the H.264 decoder to populate the new generic video coding info structures. It handles allocation of the side data buffer, collection of modes/MVs/refs for all macroblock types, and attach the final side data buffer to the output frame.

This should serve as a template for adding support for other codecs down the line.

Signed-off-by: Timothee Regaud <timothee.informatique@regaud-chapuy.fr>