configure: treat unrecognized flags as errors on MSVC
This is important for feature checking to work correctly.

It can happen that a unrecognized flag is passing compile test with a
warning only, while failing in preprocessor only mode with an error.
This makes all test_cpp calls fail and silently produces arguably broken
MSVC builds. Also all check_* doesn't work as expected, because they
think check passed, while there was an warning.

Additionally this make it in-line with GCC/Clang based builds. To test
for unrecognized and bail out early, instead of silently doing whole
build only producing warnings in log.

The /options:strict option is available starting in Visual Studio 2022
version 17.0. Because of that we cannot use check_cflags alone, it would
just also add this flag for older MSVC and produce warnings. So we need
to manually version check. Kind of chicken and egg problem.

Do the check before adding extra flags from the user, to ensure that we
don't fail silently preprocessor check, with invalid flags on older
MSVC. Note that it behaves differently depending if we are compiling or
preprocessing only.

Use check_cflags even after version check, for non-MSVC compilers that
might impersonate MSVC, but not have support for this flag. Clang-CL
sets _MSC_FULL_VER as one example.

Signed-off-by: Kacper Michajłow <kasper93@gmail.com>