avfilter/vf_scale: don't ignore incoming chroma location
This filter was, for some reason, always ignoring the incoming chroma
location in favor of the user-specified value, even when that value was set
to the default (unspecified).

This has been the status quo for quite some time, although commit 04ce01df0bb
made the situation worse by changing it from only happening when scaling is
needed, to always happening even in the no-op case.