vf_colorspace: Add an option to clamp trc LUT output
Add a new flag to the vf_colorspace filter which provides the user an
option to clamp the linear and delinear transfer characteristics LUT
values to the [0, 1] represented range. This helps constrain the
potential value range when converting between colorspaces.

Sorry about sending this multiple times, my comments outside my commit
message kept being removed from the patch for some reason. I was trying
to respond to the comment on confirming no additional time added.

I ran "time" with a command running just this filter on YUV frames and
confirmed that no significant time was adding with this patch. They both
averaged to the same time to the hundredth second.

Signed-off-by: Drew Dunne <asdunne@google.com>