RFC: Add demuxer command API (PR #20349)
PR #20349 opened by Marvin Scholz (ePirat)
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20349
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20349.patch

This PR is a proposal for a new API in `libavformat`, which I call the command API, it adds the ability to send commands to demuxers and receive replies for these commands as well. Depending on the demuxer and the command, it could also be possible to just receive a command reply or just send a command without getting a reply, this would be up to the command id documentation to specify.

It adds two new APIs to libavformat:

- `avformat_send_command`
- `avformat_receive_command_reply`

Both functions need to be passed the command ID which identifies the command to send or for which to receive a reply. It also uses void pointers for the data payloads, for lack of any better way of generic data handling in C.

To illustrate how this would be used I added a simple example for RTSP's `SET_PARAMETER` command, note that this implementation right now is not fully done in this PR and should not yet be the focus of reviews.

>From a99a5e3fcde0758a57ad2fd0bcfad77d2686439c Mon Sep 17 00:00:00 2001
From: Marvin Scholz <epirat07@gmail.com>
Date: Wed, 27 Aug 2025 03:07:00 +0200
Subject: [PATCH 1/3] avformat: Add new demuxer command API

This new API adds the ability to send commands to the demuxer and also
receive replies to these commands.

It is useful in cases like the RTSP demuxer, where it might be desirable
for the API user to send commands like SET_PARAMETER or PLAY and PAUSE.

Sending and receiving are decoupled, as to not require waiting for a
reply when sending, as with some demuxers and depending on the command
it could be necessary to process further packets in the meantime
before requesting a reply.