avformat/udp: Select output interfaces for ipv6 multicast
This fixes two old TODO's in ipv6 multicast handling.
If the system as SIOCGIFINDEX ioctl a helper function to iterate over
all interfaces added to join the multicast group on interface if approperite.

The default value is IN6ADDR_ANY (::) and it then joins all interfaces
that are up and have multicast support. Local address can be
specified and then it join ALL interfaces that use that specific
local address.

Limitations (TODO's)
Handling when network configuration is changed. if up/down etc.

Change-Id: Ifea9c0504469b108de94643a525d9b55fad5466f
Signed-off-by: Peter Enderborg <peterend@axis.com>