libavfilter/vf_remap_opencl: make kernel handle subsampled chroma
This patch fixes handling YUV frames with subsampled chroma.
Before this patch, the remap kernel was already aware of multiple planes,
but it treated the subsampled chroma incorrectly. The maps were read
based on destination coordinates and the resulting source coordinates
were also not correctly transformed back to the source image.

Also, for both Y and UV, the wrong vector components were used.

With this patch, it is possible to have a decode->remap->encode
chain entirely on the GPU. I'm using it to reframe videos from an
insta360 camera. On a Lenovo T480 with the built-in UHD 620 GPU, it
can achieve about 5x real-time (VAAPI decode -> remap_opencl
-> VAAPI encode).

Signed-off-by: Matthias Welwarsky <ffmpeg@welwarsky.de>