Fix incorrect extraction of Origin (PR #20656)
PR #20656 opened by anders-mjoll
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20656
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20656.patch

The code that attempted to extract origin (4B.02) was put in the wrong
metadata handler and was also read as the wrong integer type. According
ST 377-1 Origin is of type "Position" (Int64) in the Timeline Track, not
a 8 bit unsigned integer in the Sequence.

>From a958b6b8ef9e102b4846fd52882a9983600894ee Mon Sep 17 00:00:00 2001
From: Anders Rein <anders@onemimir.com>
Date: Mon, 6 Oct 2025 16:23:45 +0200
Subject: [PATCH] Fix incorrect extraction of Origin

The code that attempted to extract origin (4B.02) was put in the wrong
metadata handler and was also read as the wrong integer type. According
ST 377-1 Origin is of type "Position" (Int64) in the Timeline Track, not
a 8 bit unsigned integer in the Sequence.