WIP: Add encoder and decoder for MEBX/Metadata Boxed (PR #20775)
PR #20775 opened by Lukas (lholliger)
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20775
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20775.patch

ISO 14496-12 defines a timed metadata allowing for metadata per-frame
or sub-frame, this patch aims to implement a decoder to validate the
data in the moov atom as well as the frame data. This also implements
a basic encoder to write this data back. This also implements some
newer requirements for the dtyp and keyd atoms defined on Apple's
documentation as this format tends to currently only be used for
certain Apple videos such as subtitles on Vision Pro or Cinematic Mode
on iPhone, but this format is by no means limited to just Apple's usage.

A lot was edited here and a part of me feels like its a bit too much since
I had to go into the decode.c and others to get DATA to behave as frames,
but since this format is timed as frames was it felt necessary, but it seems
something like this hasn't been done much before so any help and teaching
is appreciated!

This is based on PR #20738 but since it changes a lot and isn't the original
idea I thought a new PR would suit this best.

>From 42d0970e24b7fe02299fde4c96fe265dbd31a136 Mon Sep 17 00:00:00 2001
From: lholliger <14064434+lholliger@users.noreply.github.com>
Date: Tue, 28 Oct 2025 09:37:13 -0400
Subject: [PATCH] Add encoder and decoder for MEBX/Metadata Boxed

ISO 14496-12 defines a timed metadata allowing for metadata per-frame
or sub-frame, this patch aims to implement a decoder to validate the
data in the moov atom as well as the frame data. This also implements
a basic encoder to write this data back. This also implements some
newer requirements for the dtyp and keyd atoms defined on Apple's
documentation as this format tends to currently only be used for
certain Apple videos such as subtitles on Vision Pro or Cinematic Mode
on iPhone, but this format is by no means limited to just Apple's usage.