avfilter/vf_libplacebo: don't override existing rotation attribute (PR #20831)
PR #20831 opened by Niklas Haas (haasn)
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20831
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20831.patch

This is a regression introduced by the addition of the rotation option,
which overrode the existing rotation attribute that may have been set to
the image.

To fix it, add the rotation istead of setting it - however we have to do this
directly when mapping, so as to not add it multiple times.

Fixes: 4f623b4c59

>From c42c4cd665d0d79b7ffe3d5a668fbccd8898c905 Mon Sep 17 00:00:00 2001
From: Niklas Haas <git@haasn.dev>
Date: Mon, 3 Nov 2025 18:18:33 +0100
Subject: [PATCH 1/2] avfilter/vf_libplacebo: don't override existing rotation
 attribute

This is a regression introduced by the addition of the rotation option,
which overrode the existing rotation attribute that may have been set to
the image.

To fix it, add the rotation istead of setting it - however we have to do this
directly when mapping, so as to not add it multiple times.

Fixes: 4f623b4c59c3c838f588e9def7b59dcc26dee7b9