avcodec: Explicitly use a void cast instead of av_unused in get_bits.h
Currently, av_unused doesn't work when building with MSVC.
(In the future, if building in C23 mode, it could be possible
that it would work using the [[maybe_unused]] attribute though.)

Convert this specific case of silencing an unused variable to
a void cast; this silences almost 400 warnings in builds with
MSVC, getting much better signal-to-noise ratio in the warnings
there.

After this, a build with MSVC still has almost 300 warnings, but
most of the remaining ones actually do point out inconsistencies
that would be good to fix.