avcodec/psd: Support auxiliary channels (PR #20966)
PR #20966 opened by jchw
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20966
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/20966.patch

Photoshop documents can contain additional "auxiliary" channels that don't take part in layer compositing, therefore, using the channel count to determine the presence of an alpha channel in the merged image is incorrect. Instead, as per the PSD specification, use the sign of the layer count (present in the layers and masks section) to determine if there is an alpha channel, then determine the number of "primary" channels using both the presence of the alpha channel and the pixel format.

(And yes, this really is the correct way to handle this; it's [documented here](https://www.adobe.com/devnet-apps/photoshop/fileformatashtml/#50577409_pgfId-1031423) and I manually verified this behavior going back to at least Photoshop 4.0.1.)

I've attached a new FATE sample (`psd/lena-rgbxx.psd`) that tests this, an rgb image (no alpha channel) with two extra channels. I'm new to contributing here and it seems like the contributing guide is a bit out of date so please bear with me, I wasn't sure if I should've emailed the FATE sample ahead of time or not.

>From 8e7d560c61910abc2ce6a4a1c9aad1302ddd94c9 Mon Sep 17 00:00:00 2001
From: John Chadwick <john@jchw.io>
Date: Tue, 18 Nov 2025 23:45:34 -0500
Subject: [PATCH 1/3] avcodec/psd: Support auxiliary channels

Photoshop documents can contain additional "auxiliary" channels that
don't take part in layer compositing, therefore, using the channel count
to determine the presence of an alpha channel in the merged image is
incorrect. Instead, as per the PSD specification, use the sign of the
layer count (present in the layers and masks section) to determine if
there is an alpha channel, then determine the number of primary channels
using both the presence of the alpha channel and the pixel format.