avutil/opt: support DOS paths in key value parser
This allow to pass DOS path in filter graph string. Although we use `:`
as separator, we can make an exception for DOS path looking string. In
which cases first `:` is allowed to be included in the value.

Fixes drawvg filter, which would otherwise fail to parse `file` option
when DOS path is provided.

Fixes: fate-filter-drawvg-video