hwcontext_vulkan: add support for implict DRM sync for export (PR #21042)
PR #21042 opened by russelltg
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/21042
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/21042.patch

When a frame is exported to DRM, it may be written to to read to in an asyncronous fashion. Make sure, on unmap of a Vulkan frame that was mapped to DRM, to import any fences that were put on the dmabuf.

The context for myself here is I'm working on a vulkan-video based screen recorder, which allocates vulkan frames using ffmpeg, then exports them as DRM dmabuf to send to the compositor, which then writes to it.

Without this change, I see all sorts of frame corruption because the read from ffmpeg's vulkan filter happens before/during the compositor writing to it.

>From 52c81b688fcc77f6274c64911c67ace14cb7f4ec Mon Sep 17 00:00:00 2001
From: Russell Greene <russellgreene8@gmail.com>
Date: Fri, 28 Nov 2025 23:04:57 -0700
Subject: [PATCH] hwcontext_vulkan: add support for implict DRM sync for export

When a frame is exported to DRM, it may be written to to read to in an asyncronous fashion. Make sure, on unmap of a Vulkan frame that was mapped to DRM, to import any fences that were put on the dmabuf