avformat/hlsenc: fix IVs only be auto generated once (PR #21088)
PR #21088 opened by Jack Lau (JackLau)
URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/21088
Patch URL: https://code.ffmpeg.org/FFmpeg/FFmpeg/pulls/21088.patch

Fix #20970

In previous code, it generates IV for first segment
by its sequence number when the IV don't be provided.
But it's wrong that use the first generated IV for
every segments.

Refer to RFC 8216 section 5.2 and 4.3.2.4,

- EXT-X-KEY tag with a KEYFORMAT of "identity" that
does not have an IV attribute indicates that the
Media Sequence Number is to be used as the IV.

- KEYFORMAT attribute is OPTIONAL; its absence
indicates an implicit value of "identity".

- The EXT-X-KEY tag applies to every Media Segment
between it and the next EXT-X-KEY tag in the Playlist
file with the same KEYFORMAT attribute (or the end
of the Playlist file).

This patch adds auto_iv variable to record whether
we should auto generate IV for every segments.
If the auto_iv is enabled, this patch does these changes:
1. IVs will be auto generated for every segments
2. Omit IV info in EXT-X-KEY tag
3. Only show the EXT-X-KEY tag once for these segments

Example:

command:

ffmpeg -i input.mp4
-f hls -hls_key_info_file enc.keyinfo
-hls_segment_filename chunk_%06d.ts
-y playlist.m3u8

m3u8:

 #EXTM3U
 #EXT-X-VERSION:3
 #EXT-X-TARGETDURATION:6
 #EXT-X-MEDIA-SEQUENCE:0
 #EXT-X-KEY:METHOD=AES-128,URI="http://localhost:8000/enc.key"
 #EXTINF:4.266667,
 chunk_000000.ts
 #EXTINF:5.733333,
 chunk_000001.ts
 #EXT-X-ENDLIST

Signed-off-by: Jack Lau <jacklau1222gm@gmail.com>

>From 2ee8b1596f626d608f6fac50f18af0c5f967010e Mon Sep 17 00:00:00 2001
From: Jack Lau <jacklau1222gm@gmail.com>
Date: Wed, 3 Dec 2025 17:22:27 +0800
Subject: [PATCH] avformat/hlsenc: fix IVs only be auto generated once

Fix #20970

In previous code, it generates IV for first segment
by its sequence number when the IV don't be provided.
But it's wrong that use the first generated IV for
every segments.

Refer to RFC 8216 section 5.2 and 4.3.2.4,

- EXT-X-KEY tag with a KEYFORMAT of "identity" that
does not have an IV attribute indicates that the
Media Sequence Number is to be used as the IV.

- KEYFORMAT attribute is OPTIONAL; its absence
indicates an implicit value of "identity".

- The EXT-X-KEY tag applies to every Media Segment
between it and the next EXT-X-KEY tag in the Playlist
file with the same KEYFORMAT attribute (or the end
of the Playlist file).

This patch adds auto_iv variable to record whether
we should auto generate IV for every segments.
If the auto_iv is enabled, this patch does these changes:
1. IVs will be auto generated for every segments
2. Omit IV info in EXT-X-KEY tag
3. Only show the EXT-X-KEY tag once for these segments

Example:

command:

ffmpeg -i input.mp4
-f hls -hls_key_info_file enc.keyinfo
-hls_segment_filename chunk_%06d.ts
-y playlist.m3u8

m3u8:

 #EXTM3U
 #EXT-X-VERSION:3
 #EXT-X-TARGETDURATION:6
 #EXT-X-MEDIA-SEQUENCE:0
 #EXT-X-KEY:METHOD=AES-128,URI="http://localhost:8000/enc.key"
 #EXTINF:4.266667,
 chunk_000000.ts
 #EXTINF:5.733333,
 chunk_000001.ts
 #EXT-X-ENDLIST

Signed-off-by: Jack Lau <jacklau1222gm@gmail.com>